requirements.txt
numpy==1.26.4
opencv-python-headless==4.10.0.84
paddleocr
fastapi
uvicorn[standard]
pillow
python-multipart
openai
scipy